Add python2.7 support
- Ensure python 3.5 can use async client
- Include README in package dist
